backend/pyproject.toml
[project]
name = "fuse-home-backend"
version = "0.1.0"
description = "Multi-MCP Server with Google Gen AI Integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.0",
    "google-genai",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "gradio",
    "aiohttp>=3.8.0",
    "setuptools>=80.9.0",
    "fastapi[all]>=0.117.1",
]


[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
